requirements.txt
Django
gunicorn
djangorestframework
djangorestframework-simplejwt
django-filter
